requirements.txt
streamlit>=1.25.0
pandas>=1.5
numpy>=1.23
matplotlib>=3.5
plotly>=5.6
pytest>=7.0
python-dateutil
openpyxl>=3.1